scripts/requirements.txt
# Python dependencies for Agent Factory scripts
#
# Install with: pip install -r requirements.txt

# Flashcard export to Anki
genanki>=2.1.0

# Web scraping for curriculum sync
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Search index building (optional - uses built-in json)
# No additional dependencies needed

# Data validation and schema
pyyaml>=6.0

# Progress bars for long operations (optional)
tqdm>=4.66.0
